app/requirements.txt
fastapi==0.115.6
psycopg2==2.9.10
pydantic==2.10.5
python-dotenv==1.0.1
uvicorn==0.34.0
